refactor(quire): split the wrapper types into core / manipulators / iostream (#1334)
Phase 2 group 5: quire, interval and faithful -- the three types that wrap or are generic
over another number system rather than defining an encoding of their own.

                umbrella before        core after
    quire       81,392 lines / 5 I/O   72,127 / 0
    interval    71,938 / 5             55,016 / 0
    faithful    65,176 / 5             46,988 / 0

These were expected to be cheap because they are templates over a Scalar and pull in no
concrete number system. Two of the three had something in the way.

quire's impl was ALREADY at 0 I/O -- its blockbinary and blocktriple substrates were
cleaned in Phase N -- yet the core would not compile. quire_properties(), a report builder
that formats the accumulator's configuration into a std::stringstream, sat at the top of
the arithmetic impl and had never included <sstream>: it worked only behind a prior
include. It is a text function, so it moved to manipulators.hpp.

faithful's manipulators.hpp is dead code. It was commented out of the umbrella long before
this epic and does not compile -- 18 errors on main. Moving components() there would have
stranded it, so components() went to iostream.hpp, which the umbrella does include, and
manipulators.hpp is left byte-identical to main rather than half-fixed. The umbrella now
says why. faithful's two std::cerr "TBD" sites became fprintf, and its impl re-points from
native/ieee754.hpp to ieee754_core.hpp + manipulators_core.hpp.

interval was the only genuinely clean one: operator<</operator>> to iostream.hpp,
to_string/type_tag to manipulators.hpp, nothing else.

quire_impl.hpp's operator<< friend declaration takes <iosfwd>, deliberately NOT <ostream>
as the IWYU sweep suggests -- a declaration needs only the forward header, and <ostream>
would put a stream header straight back into the core.

Closed 21 IWYU gaps across the three subtrees. The only ones left are in faithful's dead
manipulators.hpp, left alone on purpose.

Verified: all three core gates pass on gcc and clang with #error tripwires on the stream
include guards; 191 consumers build and run clean on both compilers. 53 more fail, every
one identical on main and none related: 45 in archive/ referencing posit1/posito/valid
removed in #1243, 7 needing blas/blas.hpp and 1 needing type_dispatch.hpp -- files that
want extra -I paths a flat sweep does not supply. The expansion_algebra tutorial still
times out; it loops forever on EOF, which predates this work.

Part of #1334.
